move helpers next to DJPP
