+++ src/main/java/spoon/reflect/visitor/PrinterHelper.java
@@ -14,7 +14,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-C license and that you accept its terms.
  */
-package spoon.reflect.visitor.printer;
+package spoon.reflect.visitor;
 
 import spoon.compiler.Environment;
 import spoon.reflect.code.BinaryOperatorKind;
